made index sizing use int()

+++ test/exampleOptimizationAEP-Rotor.py
@@ -116,7 +116,7 @@ if __name__ == "__main__":
         prob.driver.add_desvar('yaw%i' % direction_id, lower=-30.0, upper=30.0, scaler=1E-1)
 
     # add constraints
-    prob.driver.add_constraint('sc', lower=np.zeros(((nTurbs-1.)*nTurbs/2.)))
+    prob.driver.add_constraint('sc', lower=np.zeros(int(((nTurbs-1.)*nTurbs/2.))))
 
     # set voi_red options
     # prob.root.ln_solver.options['single_voi_relevance_reduction'] = True
